Got robot code to upload
Still doesn't work. It just barrages the riolog with a bunch of errors. The ones I could pick out is: CAN too stale, check wires (we might've fixed that); make sure hardware is updated to v6 (I think it might be a phoenix thing); and make sure the controller is plugged in (I'm not sure what the deal with that is because the joysticks are responding).